Fixed ts-project-with-jest
A second take on ts-project-with-jest, this time with different settings in package.json(this change might be redundant) and tsconfig.json.

the code we add to tsconfig on page 8 has been left out due to it causing .ts to become an unrecognised file type.

The primary difference is that this version works :)