Readme updates and start up, vehicle tweaks
Added more content to the readme's program flow, vehicles and future additions sections.
fixed a typo in the vehicle type where position was spelt postion( that cascaded to a fair few changes)
Added a type to specify the available vehicle types.
updated tests for invalid vic types on move functions since the old placeholder of "boat" is no longer allowed.
added call to switchVics() to startup() so that the starter rover is actually selected.